fix(zellij-utils): Fix duplication of `PaneId` in protobuf
