+++ zellij-utils/assets/prost/api.event.rs
@@ -102,7 +102,7 @@ pub struct ClientInfo {
     #[prost(uint32, tag = "1")]
     pub client_id: u32,
     #[prost(message, optional, tag = "2")]
-    pub pane_id: ::core::option::Option<PaneId>,
+    pub pane_id: ::core::option::Option<super::pane_id::PaneId>,
     #[prost(string, tag = "3")]
     pub running_command: ::prost::alloc::string::String,
     #[prost(bool, tag = "4")]
@@ -126,16 +126,7 @@ pub struct CommandPaneReRunPayload {
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct PaneClosedPayload {
     #[prost(message, optional, tag = "1")]
-    pub pane_id: ::core::option::Option<PaneId>,
-}
-/// duplicate of plugin_command.PaneId because protobuffs don't like recursive imports
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct PaneId {
-    #[prost(enumeration = "PaneType", tag = "1")]
-    pub pane_type: i32,
-    #[prost(uint32, tag = "2")]
-    pub id: u32,
+    pub pane_id: ::core::option::Option<super::pane_id::PaneId>,
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
@@ -597,33 +588,6 @@ impl EventType {
             "FailedToChangeHostFolder" => Some(Self::FailedToChangeHostFolder),
             "PastedText" => Some(Self::PastedText),
             "ConfigWasWrittenToDisk" => Some(Self::ConfigWasWrittenToDisk),
-            _ => None,
-        }
-    }
-}
-/// duplicate of plugin_command.PaneType because protobuffs don't like recursive imports
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
-#[repr(i32)]
-pub enum PaneType {
-    Terminal = 0,
-    Plugin = 1,
-}
-impl PaneType {
-    /// String value of the enum field names used in the ProtoBuf definition.
-    ///
-    /// The values are not transformed in any way and thus are considered stable
-    /// (if the ProtoBuf definition does not change) and safe for programmatic use.
-    pub fn as_str_name(&self) -> &'static str {
-        match self {
-            PaneType::Terminal => "Terminal",
-            PaneType::Plugin => "Plugin",
-        }
-    }
-    /// Creates an enum from field names used in the ProtoBuf definition.
-    pub fn from_str_name(value: &str) -> ::core::option::Option<Self> {
-        match value {
-            "Terminal" => Some(Self::Terminal),
-            "Plugin" => Some(Self::Plugin),
             _ => None,
         }
     }
